+++ packages/instrumentation-neo4j/src/types.ts
@@ -1,11 +1,12 @@
 import { Span } from '@opentelemetry/api';
 import { InstrumentationConfig } from '@opentelemetry/instrumentation';
+import type { QueryResult } from 'neo4j-driver';
 
-export type SequelizeResponseCustomAttributesFunction = (span: Span, response: any) => void;
+export type Neo4jResponseCustomAttributesFunction = (span: Span, response: QueryResult) => void;
 
-export interface SequelizeInstrumentationConfig extends InstrumentationConfig {
+export interface Neo4jInstrumentationConfig extends InstrumentationConfig {
     /** hook for adding custom attributes using the response payload */
-    responseHook?: SequelizeResponseCustomAttributesFunction;
+    responseHook?: Neo4jResponseCustomAttributesFunction;
     /** Set to true if you only want to trace operation which has parent spans */
     ignoreOrphanedSpans?: boolean;
     /**