+++ packages/instrumentation-typeorm/src/types.ts
@@ -6,4 +6,9 @@ export type TypeormResponseCustomAttributesFunction = (span: Span, response: any
 export interface TypeormInstrumentationConfig extends InstrumentationConfig {
     /** hook for adding custom attributes using the response payload */
     responseHook?: TypeormResponseCustomAttributesFunction;
+    /**
+     * If passed, a span attribute will be added to all spans with key of the provided "moduleVersionAttributeName"
+     * and value of the module version.
+     */
+    moduleVersionAttributeName?: string;
 }